Changes in vector
Finished push_back() logic, adding the reserve() function which checks if the given size clips the maximum size (throws exception with noreturn), or if it's less than the actual capacity, which returns void. Else, it will alloc new memory with the new capacity, destroy the old memory and dealloc, and then assign each pointer correctly. Push back checks the capacity and calculates a new one based on this logic:
    - Base cases: ++size if n = 0 or n = 1.
    - size + (size / 2) if n > 1
This way we get quite a few allocs for the first pushes but as we push more data into the vector there'll be less allocs while we keep the capacity at a reasonable size. There's also a change in memory operators. The header caused a lot of errors when using the official STL, beacuse the operators were arleady declared. Defining the macro _REDEFINE_NEW_OPERATORS_ will use my implementation of the operators.