Pair struct, and new macros
New macros in cxx_config, including _ILIBCXX_CONSTEXPR which automatically checks for the current c++ version to write constexpr or not. Then, __CXX_HASCXX0X__ which is a short way of checking for 0X or 03 mode.

The pair struct is now part of my STL. I decided to remove part of the original STL constructors as it involded weird type_traits structs which I don't even understand. It was mostly for rvalue forwarding so it's not *that* useful.