Changes to vector, common_types and hash.
Added more constructors to vector, which serve to initialize with a given value. Fixed naming in _Vector_base (_Alloc and function parameter _Alloc caused errors in compilation time).

Added file common_types.h which includes just typedefs for all integer sizes. The purpose of this file is to get rid of __int64 which is only Windows-compatible.

Added file hash.h which implements (for now) the xxhash function.